+++ examples/echo-bot/src/Clients.ts
@@ -2,10 +2,8 @@ import * as Console from "@mmlngl/effect-messagekit-provider-console";
 import * as Line from "@mmlngl/effect-messagekit-provider-line";
 import * as Layer from "effect/Layer";
 
-const LineClientLayer = Line.Client.LineClient.layer.pipe(
-  Layer.provide(Line.Config.LineConfig.layerFromEnv),
-);
+const LineClientLayer = Line.Client.layer;
 
-const ConsoleClientLayer = Console.Client.ConsoleClient.layer;
+const ConsoleClientLayer = Console.Client.layer;
 
 export const ClientsLayer = Layer.mergeAll(LineClientLayer, ConsoleClientLayer);