+++ packages/core/src/index.ts
@@ -1,2 +1,0 @@
-export * as Domain from "./domain";
-export * as Infra from "./infra";