+++ packages/core/tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: ["src/index.ts"],
+  entry: ["src/domain/index.ts", "src/infra/index.ts"],
   format: ["esm"],
   dts: true,
   sourcemap: true,