+++ packages/provider-console/src/Messages/index.ts
@@ -1,3 +1,13 @@
+import * as Schema from "effect/Schema";
+import { ConsoleJsonMessage } from "./json-message";
+import { ConsoleMarkdownMessage } from "./markdown-message";
+
+export const AnyConsoleMessage = Schema.Union(
+  ConsoleJsonMessage,
+  ConsoleMarkdownMessage,
+);
+
+export type AnyConsoleMessage = typeof AnyConsoleMessage.Type;
+
 export * as JsonMessage from "./json-message";
-export * from "./Messages";
 export * as MarkdownMessage from "./markdown-message";