+++ packages/provider-line/src/Messages/flex-message.ts
@@ -1,8 +1,5 @@
 import type * as Line from "@line/bot-sdk";
-import * as Core from "@mmlngl/effect-messagekit-core";
-import * as Effect from "effect/Effect";
 import * as Schema from "effect/Schema";
-import * as Common from "./common";
 
 export const LineFlexMessageContainerBubble = Schema.Struct({
   type: Schema.Literal("bubble"),
@@ -29,7 +26,6 @@ export const LineFlexMessageContainer = Schema.Union(
 );
 
 export const LineFlexMessage = Schema.TaggedStruct("LineFlexMessage", {
-  ...Common.fields,
   type: Schema.Literal("flex"),
   altText: Schema.NonEmptyTrimmedString,
   contents: LineFlexMessageContainer,
@@ -39,25 +35,3 @@ export type LineFlexMessageType = typeof LineFlexMessage.Type;
 
 // Compile-time type checks to ensure compatibility with LINE SDK
 const _assert: Line.messagingApi.FlexMessage = {} as LineFlexMessageType;
-
-export const presenter: Core.Domain.Presenter.PresenterTrait<LineFlexMessageType> =
-  {
-    build: (messages) =>
-      Effect.all(
-        messages.map((msg) =>
-          Schema.decodeUnknown(LineFlexMessage)({
-            _tag: "LineFlexMessage",
-            type: "flex",
-            to: msg.recipient,
-            altText: "Flex message",
-            contents: LineFlexMessageContainerBubble.make({
-              type: "bubble",
-            }),
-          }),
-        ),
-      ).pipe(
-        Effect.mapError(
-          (cause) => new Core.Domain.Presenter.PresentBuildError({ cause }),
-        ),
-      ),
-  };