+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mmlngl/effect-messagekit-core
 
+## 0.0.2
+
+### Patch Changes
+
+- removed presenters
+
 ## 0.0.1
 
 ### Patch Changes