+++ packages/provider-line/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mmlngl/effect-messagekit-provider-line
 
+## 0.0.2
+
+### Patch Changes
+
+- removed presenters
+- Updated dependencies
+  - @mmlngl/effect-messagekit-core@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes